requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.8
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-filter==2.2.0
django-formtools==2.2
django-import-export==2.0.1
django-redis==4.11.0
django-reversion==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
Faker==4.0.0
future==0.18.2
httplib2==0.9.2
idna==2.8
importlib-metadata==1.4.0
jdcal==1.4.1
kombu==4.6.7
lxml==4.4.2
MarkupPy==1.14
more-itertools==8.1.0
odfpy==1.4.0
openpyxl==3.0.3
Pillow==7.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
qcloudsms-py==0.1.4
redis==3.3.11
requests==2.22.0
six==1.14.0
some-package==0.1
sqlparse==0.3.0
tablib==0.14.0
text-unidecode==1.3
urllib3==1.25.7
vine==1.3.0
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==1.0.0
